+++ WebIluminaMVC/Controllers/NoticesController.cs
@@ -19,28 +19,39 @@ namespace WebIluminaMVC.Controllers
         // GET: Notices
         public ActionResult Index()
         {
-            return View(db.Notice.ToList());
+            if(DataUtil.Validation())
+                return View(db.Notice.ToList());
+            else
+                return RedirectToAction("Login", "Home");
         }
 
         // GET: Notices/Details/5
         public ActionResult Details(int? id)
         {
-            if (id == null)
+            if (DataUtil.Validation())
             {
-                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                if (id == null)
+                {
+                    return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                }
+                Notice notice = db.Notice.Find(id);
+                if (notice == null)
+                {
+                    return HttpNotFound();
+                }
+                return View(notice);
             }
-            Notice notice = db.Notice.Find(id);
-            if (notice == null)
-            {
-                return HttpNotFound();
-            }
-            return View(notice);
+            else
+                return RedirectToAction("Login", "Home");
         }
 
         // GET: Notices/Create
         public ActionResult Create()
         {
-            return View();
+            if(DataUtil.Validation())
+                return View();
+            else
+                return RedirectToAction("Login", "Home");
         }
 
         // POST: Notices/Create
@@ -49,50 +60,60 @@ namespace WebIluminaMVC.Controllers
         [HttpPost]
         [ValidateAntiForgeryToken]
         public ActionResult Create([Bind(Include = "noticeID,title,publishDate,author,contents,imageUrl,active,createDate,createUser,updateDate,updateUser")] Notice notice)
-        {            
-            if (ModelState.IsValid)
+        {
+            if (DataUtil.Validation())
             {
-                notice.publishDate = DateTime.ParseExact(Request.Form["CtrlPublishDate"], "dd/MM/yyyy", null);
-                
-
-                if (Request.Files.Count > 0)
+                if (ModelState.IsValid)
                 {
-                    var file = Request.Files[0];
+                    notice.publishDate = DateTime.ParseExact(Request.Form["CtrlPublishDate"], "dd/MM/yyyy", null);
 
-                    if (file != null && file.ContentLength > 0)
+
+                    if (Request.Files.Count > 0)
                     {
-                        var fileName = Path.GetFileName(file.FileName);
+                        var file = Request.Files[0];
 
-                        notice.imageUrl = fileName;
-                        db.Notice.Add(notice);
-                        db.SaveChanges();                        
+                        if (file != null && file.ContentLength > 0)
+                        {
+                            var fileName = Path.GetFileName(file.FileName);
 
-                        Directory.CreateDirectory(Server.MapPath("~"+System.Configuration.ConfigurationManager.AppSettings["RouteImageNotice"] + notice.noticeID));
-                        var path = Path.Combine(Server.MapPath("~"+System.Configuration.ConfigurationManager.AppSettings["RouteImageNotice"] + notice.noticeID), fileName);                        
-                        file.SaveAs(path);
+                            notice.imageUrl = fileName;
+                            db.Notice.Add(notice);
+                            db.SaveChanges();
+
+                            Directory.CreateDirectory(Server.MapPath("~" + System.Configuration.ConfigurationManager.AppSettings["RouteImageNotice"] + notice.noticeID));
+                            var path = Path.Combine(Server.MapPath("~" + System.Configuration.ConfigurationManager.AppSettings["RouteImageNotice"] + notice.noticeID), fileName);
+                            file.SaveAs(path);
+                        }
                     }
+
+
+                    return RedirectToAction("Index");
                 }
 
-
-                return RedirectToAction("Index");
+                return View(notice);
             }
-
-            return View(notice);
+            else
+                return RedirectToAction("Login", "Home");
         }
 
         // GET: Notices/Edit/5
         public ActionResult Edit(int? id)
         {
-            if (id == null)
+            if (DataUtil.Validation())
             {
-                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                if (id == null)
+                {
+                    return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                }
+                Notice notice = db.Notice.Find(id);
+                if (notice == null)
+                {
+                    return HttpNotFound();
+                }
+                return View(notice);
             }
-            Notice notice = db.Notice.Find(id);
-            if (notice == null)
-            {
-                return HttpNotFound();
-            }
-            return View(notice);
+            else
+                return RedirectToAction("Login", "Home");
         }
 
         // POST: Notices/Edit/5
@@ -103,51 +124,61 @@ namespace WebIluminaMVC.Controllers
         [HttpPost]
         public ActionResult Edit(Notice notice)
         {
-            if (ModelState.IsValid)
-            {   
-                notice.publishDate = DateTime.ParseExact(Request.Form["CtrlPublishDate"], "dd/MM/yyyy", null);
-
-                if (Request.Files.Count > 0)
+            if (DataUtil.Validation())
+            {
+                if (ModelState.IsValid)
                 {
-                    var file = Request.Files[0];
+                    notice.publishDate = DateTime.ParseExact(Request.Form["CtrlPublishDate"], "dd/MM/yyyy", null);
 
-                    if (file != null && file.ContentLength > 0)
+                    if (Request.Files.Count > 0)
                     {
-                        var fileName = Path.GetFileName(file.FileName);                        
+                        var file = Request.Files[0];
 
-                        if (!Directory.Exists(Server.MapPath("~" + System.Configuration.ConfigurationManager.AppSettings["RouteImageNotice"] + notice.noticeID)))
+                        if (file != null && file.ContentLength > 0)
                         {
-                            Directory.CreateDirectory(Server.MapPath("~" + System.Configuration.ConfigurationManager.AppSettings["RouteImageNotice"] + notice.noticeID));
+                            var fileName = Path.GetFileName(file.FileName);
+
+                            if (!Directory.Exists(Server.MapPath("~" + System.Configuration.ConfigurationManager.AppSettings["RouteImageNotice"] + notice.noticeID)))
+                            {
+                                Directory.CreateDirectory(Server.MapPath("~" + System.Configuration.ConfigurationManager.AppSettings["RouteImageNotice"] + notice.noticeID));
+                            }
+                            else
+                            {
+                                System.IO.File.Delete(Server.MapPath("~" + System.Configuration.ConfigurationManager.AppSettings["RouteImageNotice"] + notice.noticeID + notice.imageUrl));
+                            }
+                            var path = Path.Combine(Server.MapPath("~" + System.Configuration.ConfigurationManager.AppSettings["RouteImageNotice"] + notice.noticeID), fileName);
+                            file.SaveAs(path);
+                            notice.imageUrl = fileName;
                         }
-                        else
-                        {
-                            System.IO.File.Delete(Server.MapPath("~" + System.Configuration.ConfigurationManager.AppSettings["RouteImageNotice"] + notice.noticeID + notice.imageUrl));
-                        }
-                        var path = Path.Combine(Server.MapPath("~" + System.Configuration.ConfigurationManager.AppSettings["RouteImageNotice"] + notice.noticeID), fileName);
-                        file.SaveAs(path);
-                        notice.imageUrl = fileName;
                     }
-                }
 
-                db.Entry(notice).State = EntityState.Modified;
-                db.SaveChanges();
+                    db.Entry(notice).State = EntityState.Modified;
+                    db.SaveChanges();
+                }
+                return View(notice);
             }
-            return View(notice);
+            else
+                return RedirectToAction("Login", "Home");
         }
 
         // GET: Notices/Delete/5
         public ActionResult Delete(int? id)
         {
-            if (id == null)
+            if (DataUtil.Validation())
             {
-                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                if (id == null)
+                {
+                    return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                }
+                Notice notice = db.Notice.Find(id);
+                if (notice == null)
+                {
+                    return HttpNotFound();
+                }
+                return View(notice);
             }
-            Notice notice = db.Notice.Find(id);
-            if (notice == null)
-            {
-                return HttpNotFound();
-            }
-            return View(notice);
+            else
+                return RedirectToAction("Login", "Home");
         }
 
         // POST: Notices/Delete/5
@@ -155,10 +186,15 @@ namespace WebIluminaMVC.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult DeleteConfirmed(int id)
         {
-            Notice notice = db.Notice.Find(id);
-            db.Notice.Remove(notice);
-            db.SaveChanges();
-            return RedirectToAction("Index");
+            if (DataUtil.Validation())
+            {
+                Notice notice = db.Notice.Find(id);
+                db.Notice.Remove(notice);
+                db.SaveChanges();
+                return RedirectToAction("Index");
+            }
+            else
+                return RedirectToAction("Login", "Home");
         }
 
         protected override void Dispose(bool disposing)
